move getMetadata + models to Models
