+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Models/APIViewFeedbackItem.cs
@@ -1,70 +1,100 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace Azure.Sdk.Tools.Cli.Helpers;
+using System.Text.Json.Serialization;
 
+namespace Azure.Sdk.Tools.Cli.Models;
 
 /// <summary>
-/// Feedback input from APIView review comments
+/// Represents complete metadata for an APIView review
 /// </summary>
-public class APIViewFeedbackItem : IFeedbackItem
+public class ReviewMetadata
 {
-    private readonly string _apiViewUrl;
-    private readonly IAPIViewFeedbackHelper _helper;
-    private readonly ILogger<APIViewFeedbackItem> _logger;
+    [JsonPropertyName("reviewId")]
+    public string ReviewId { get; set; } = string.Empty;
+    
+    [JsonPropertyName("packageName")]
+    public string PackageName { get; set; } = string.Empty;
+    
+    [JsonPropertyName("language")]
+    public string Language { get; set; } = string.Empty;
+    
+    [JsonPropertyName("revision")]
+    public RevisionMetadata? Revision { get; set; }
+}
 
-    public APIViewFeedbackItem(
-        string apiViewUrl,
-        IAPIViewFeedbackHelper helper,
-        ILogger<APIViewFeedbackItem> logger)
-    {
-        _apiViewUrl = apiViewUrl;
-        _helper = helper;
-        _logger = logger;
-    }
+/// <summary>
+/// Represents revision-specific metadata from APIView
+/// </summary>
+public class RevisionMetadata
+{
+    [JsonPropertyName("revisionId")]
+    public string RevisionId { get; set; } = string.Empty;
+    
+    [JsonPropertyName("pullRequestNo")]
+    public int? PullRequestNo { get; set; }
+    
+    [JsonPropertyName("pullRequestRepository")]
+    public string? PullRequestRepository { get; set; }
+    
+    [JsonPropertyName("revisionLabel")]
+    public string? RevisionLabel { get; set; }
+}
 
-    public async Task<FeedbackContext> PreprocessAsync(CancellationToken ct = default)
-    {
-        _logger.LogInformation("Preprocessing APIView feedback from: {Url}", _apiViewUrl);
+/// <summary>
+/// Represents a consolidated comment from a discussion thread
+/// </summary>
+// TODO: Add ThreadUrl property for direct links to APIView discussion threads
+public class ConsolidatedComment
+{
+    public string ThreadId { get; set; } = string.Empty;
+    public int LineNo { get; set; }
+    
+    /// <summary>
+    /// Line identifier from APIView. Note: _lineId is Python-specific and serves as 
+    /// placeholder for fullObjectPath until fullObjectPath is added to APIView API.
+    /// </summary>
+    public string? LineId { get; set; } = string.Empty;
+    
+    public string LineText { get; set; } = string.Empty;
+    public string Comment { get; set; } = string.Empty;
+}
 
-        // Get consolidated comments
-        var comments = await _helper.GetConsolidatedComments(_apiViewUrl);
-        
-        // Get metadata
-        var metadata = await _helper.GetMetadata(_apiViewUrl);
-
-        // Convert to feedback items
-        var feedbackItems = comments.Select(c =>
-        {
-            var text = $"API Line {c.LineNo}: {c.LineId}, Code: {c.LineText.Trim()}, ReviewComment: {c.Comment}";
-            var item = new FeedbackItem
-            {
-                Text = text,
-                Context = string.Empty
-            };
-            item.FormattedPrompt = $"Id: {item.Id}\nText: {text}\nContext: ";
-            return item;
-        }).ToList();
-
-        _logger.LogInformation("Converted {Count} comments to feedback items", feedbackItems.Count);
-
-        return new FeedbackContext
-        {
-            FormattedFeedback = string.Join("\n\n", feedbackItems.Select(f => f.FormattedPrompt)),
-            FeedbackItems = feedbackItems,
-            Language = metadata.Language,
-            PackageName = metadata.PackageName,
-            InputType = "apiview",
-            Metadata = new Dictionary<string, string>
-            {
-                ["APIViewUrl"] = _apiViewUrl
-            }
-        };
-    }
-
-    private static string FormatCommentForPrompt(ConsolidatedComment comment, string itemId)
-    {
-        var text = $"API Line {comment.LineNo}: {comment.LineId}, Code: {comment.LineText.Trim()}, ReviewComment: {comment.Comment}";
-        return $"Id: {itemId}\nText: {text}\nContext: ";
-    }
+/// <summary>
+/// Represents a raw comment from APIView API
+/// </summary>
+internal class APIViewComment
+{
+    [JsonPropertyName("lineNo")]
+    public int LineNo { get; set; }
+    
+    [JsonPropertyName("_lineId")]
+    public string? LineId { get; set; }
+    
+    [JsonPropertyName("_lineText")]
+    public string? LineText { get; set; }
+    
+    [JsonPropertyName("createdOn")]
+    public string? CreatedOn { get; set; }
+    
+    [JsonPropertyName("upvotes")]
+    public int Upvotes { get; set; }
+    
+    [JsonPropertyName("downvotes")]
+    public int Downvotes { get; set; }
+    
+    [JsonPropertyName("createdBy")]
+    public string? CreatedBy { get; set; }
+    
+    [JsonPropertyName("commentText")]
+    public string? CommentText { get; set; }
+    
+    [JsonPropertyName("isResolved")]
+    public bool IsResolved { get; set; }
+    
+    [JsonPropertyName("severity")]
+    public string? Severity { get; set; }
+    
+    [JsonPropertyName("threadId")]
+    public string? ThreadId { get; set; }
 }
