+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Prompts/Templates/APIViewFeedbackIssueTemplate.cs
@@ -3,6 +3,7 @@
 
 using System.Text;
 using Azure.Sdk.Tools.Cli.Helpers;
+using Azure.Sdk.Tools.Cli.Models;
 using Azure.Sdk.Tools.Cli.Services.APIView;
 
 namespace Azure.Sdk.Tools.Cli.Prompts.Templates;
