+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Services/APIView/APIViewService.cs
@@ -4,6 +4,7 @@ public interface IAPIViewService
 {
     Task<string?> GetRevisionContent(string apiRevisionId, string reviewId, string contentReturnType);
     Task<string?> GetCommentsByRevisionAsync(string revisionId);
+    Task<string?> GetMetadata(string revisionId);
 }
 
 public class APIViewService : IAPIViewService
@@ -40,6 +41,19 @@ public class APIViewService : IAPIViewService
         {
             _logger.LogWarning("Received empty response for revisions {ActiveRevisionId}", apiRevisionId);
             return null;
+        }
+
+        return result;
+    }
+
+    public async Task<string?> GetMetadata(string revisionId)
+    {
+        string endpoint = $"/api/reviews/metadata?revisionId={revisionId}";
+        string? result = await _httpService.GetAsync(endpoint);
+
+        if (result == null)
+        {
+            _logger.LogWarning("No metadata found for revision {RevisionId}", revisionId);
         }
 
         return result;
