+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Tools/APIView/APIViewReviewTool.cs
@@ -149,7 +149,7 @@ public class APIViewReviewTool : MCPMultiCommandTool
         }
     }
 
-    private (string revisionId, string reviewId) ExtractIdsFromUrl(string url)
+    public static (string revisionId, string reviewId) ExtractIdsFromUrl(string url)
     {
         if (string.IsNullOrWhiteSpace(url))
         {
@@ -178,8 +178,7 @@ public class APIViewReviewTool : MCPMultiCommandTool
         }
         catch (Exception ex) when (ex is not ArgumentException)
         {
-            _logger.LogError(ex, "Failed to parse APIView URL {Url}", url);
-            throw new ArgumentException($"Error parsing URL: {ex.Message}", nameof(url));
+            throw new ArgumentException($"Error parsing URL: {ex.Message}", nameof(url), ex);
         }
     }
 }
